클래스명, 메서드명 변경 및 Model.kt에 Image.kt 중복 코드 리팩토링
