+++ app/src/main/java/com/example/naversearch/category/Cafe.kt
@@ -6,12 +6,12 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.naversearch.databinding.FrgCafeBinding
-import com.example.naversearch.model.TextModel
+import com.example.naversearch.model.Model
 
 class Cafe : Fragment() {
 
     private lateinit var binding: FrgCafeBinding
-    val model = TextModel()
+    val model = Model()
 
     override fun onCreateView(
         inflater: LayoutInflater,
