+++ app/src/main/java/com/example/naversearch/model/NaverAPI.kt
@@ -2,7 +2,10 @@ package com.example.naversearch.model
 
 import com.example.naversearch.ResultGetSearch
 import retrofit2.Call
-import retrofit2.http.*
+import retrofit2.http.GET
+import retrofit2.http.Headers
+import retrofit2.http.Path
+import retrofit2.http.Query
 
 interface NaverAPI {
     @Headers(
@@ -10,7 +13,7 @@ interface NaverAPI {
         "X-Naver-Client-Secret:6aM7WmBUDV"
     )
     @GET("v1/search/{id}.json")
-    fun getSearchNews(
+    fun getSearch(
         @Path("id") id: String,    // use @Path() instead of @Query()
         @Query("query") query: String,
         @Query("display") display: Int? = 9,
